moved api commands and added the ability to register with command name and callback
